pyproject.toml
[tool.poetry]
name = "load-testing"
version = "0.1.0"
description = ""
authors = ["Yuchan Lee <lasuillard@gmail.com>"]
readme = "README.md"
classifiers = ["Private :: Do not Upload"]

[tool.poetry.dependencies]
python = "^3.11"
locust = "^2.15.1"
faker = "^18.9.0"
factory-boy = "^3.2.1"
locust-plugins = "^3.3.1"
opentelemetry-api = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
opentelemetry-exporter-otlp = "^1.19.0"

[tool.poetry.group.dev.dependencies]
setuptools = "^67.7.2"
ipykernel = "^6.23.1"
ruff = "^0.0.280"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
select = ["ALL"]
ignore = ["FIX", "TD003", "ANN101", "D100", "D105", "D106"]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["D104"]
"environments/*/*.py" = ["INP001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["config"]
